# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DigiProducts API — single-tenant AI digital product creation platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: getProducts
      tags: [products]
      parameters:
        - name: type
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: ownerId
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Product list (role scoped)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Draft created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{productId}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Product detail with chapters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{productId}/duplicate:
    post:
      operationId: duplicateProduct
      tags: [products]
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
      responses:
        "201":
          description: New draft clone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{productId}/archive:
    post:
      operationId: archiveProduct
      tags: [products]
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{productId}/submit-review:
    post:
      operationId: submitForReview
      tags: [products]
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{productId}/review:
    post:
      operationId: reviewProduct
      tags: [reviews]
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDecision"
      responses:
        "200":
          description: Review recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{productId}/chapters:
    post:
      operationId: addChapter
      tags: [products]
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChapterInput"
      responses:
        "201":
          description: Chapter added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
  /products/{productId}/chapters/{chapterId}:
    patch:
      operationId: updateChapter
      tags: [products]
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
        - name: chapterId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChapterUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
    delete:
      operationId: deleteChapter
      tags: [products]
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
        - name: chapterId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /products/{productId}/comments:
    get:
      operationId: getComments
      tags: [reviews]
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: addComment
      tags: [reviews]
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /products/{productId}/sales-copy:
    get:
      operationId: getSalesCopy
      tags: [products]
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Sales copy draft (may be empty)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesCopy"
    patch:
      operationId: updateSalesCopy
      tags: [products]
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalesCopyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesCopy"
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
...
    Product:
      type: object
      required: [id, ownerId, ownerName, type, title, status, chapterCount, wordCount, createdAt, updatedAt]
      properties:
        id: { type: string }
        ownerId: { type: string }
        ownerName: { type: string }
        type: { type: string, enum: [ebook, lead_magnet, uploaded] }
        title: { type: string }
        subtitle: { type: ["string", "null"] }
        topic: { type: ["string", "null"] }
        audience: { type: ["string", "null"] }
        tone: { type: ["string", "null"] }
        language: { type: ["string", "null"] }
        depth: { type: ["string", "null"] }
        status: { type: string, enum: [draft, generating, ready, in_review, changes_requested, approved, archived] }
        coverConfig:
          type: ["object", "null"]
          additionalProperties: true
        chapterCount: { type: integer }
        wordCount: { type: integer }
        createdAt: { type: string }
        updatedAt: { type: string }
    ProductDetail:
      type: object
      required: [product, chapters]
      properties:
        product:
          $ref: "#/components/schemas/Product"
        chapters:
          type: array
          items:
            $ref: "#/components/schemas/Chapter"
        latestReview:
          $ref: "#/components/schemas/Review"
    ProductInput:
      type: object
      required: [type, title, topic]
      properties:
        type: { type: string, enum: [ebook, lead_magnet, uploaded] }
        title: { type: string, minLength: 1 }
        subtitle: { type: string }
        topic: { type: string, minLength: 1 }
        audience: { type: string }
        tone: { type: string }
        language: { type: string }
        depth: { type: string }
        chapterCount: { type: integer }
        keyPoints: { type: string }
        cta: { type: string }
        leadMagnetFormat: { type: string }
    ProductUpdate:
      type: object
      properties:
        title: { type: string }
        subtitle: { type: string }
        topic: { type: string }
        audience: { type: string }
        tone: { type: string }
        language: { type: string }
        depth: { type: string }
        coverConfig:
          type: object
          additionalProperties: true
    Chapter:
      type: object
      required: [id, productId, orderIndex, title, status, wordCount]
      properties:
        id: { type: string }
        productId: { type: string }
        orderIndex: { type: integer }
        title: { type: string }
        summary: { type: ["string", "null"] }
        contentMd: { type: ["string", "null"] }
        status: { type: string, enum: [pending, generating, ready, failed] }
        wordCount: { type: integer }
        errorMessage: { type: ["string", "null"] }
    ChapterInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
        summary: { type: string }
        contentMd: { type: string }
    ChapterUpdate:
      type: object
      properties:
        title: { type: string }
        summary: { type: string }
        contentMd: { type: string }
...
    ReviewDecision:
      type: object
      required: [decision]
      properties:
        decision: { type: string, enum: [approved, changes_requested] }
        comment: { type: string }
    Review:
      type: object
      required: [id, productId, decision, createdAt]
      properties:
        id: { type: string }
        productId: { type: string }
        reviewerName: { type: ["string", "null"] }
        decision: { type: string }
        comment: { type: ["string", "null"] }
        decidedAt: { type: ["string", "null"] }
        createdAt: { type: string }
    Comment:
      type: object
      required: [id, productId, authorId, authorName, body, createdAt]
      properties:
        id: { type: string }
        productId: { type: string }
        chapterId: { type: ["string", "null"] }
        authorId: { type: string }
        authorName: { type: string }
        body: { type: string }
        createdAt: { type: string }
    CommentInput:
      type: object
      required: [body]
      properties:
        body: { type: string, minLength: 1 }
        chapterId: { type: string }
...
    SalesCopy:
      type: object
      required: [productId]
      properties:
        productId: { type: string }
        headline: { type: ["string", "null"] }
        subheadline: { type: ["string", "null"] }
        bullets:
          type: array
          items: { type: string }
        whoItsFor: { type: ["string", "null"] }
        faq:
          type: array
          items:
            type: object
            required: [question, answer]
            properties:
              question: { type: string }
              answer: { type: string }
        ctaText: { type: ["string", "null"] }
        suggestedPriceBand: { type: ["string", "null"] }
        updatedAt: { type: ["string", "null"] }
    SalesCopyUpdate:
      type: object
      properties:
        headline: { type: string }
        subheadline: { type: string }
        bullets:
          type: array
          items: { type: string }
        whoItsFor: { type: string }
        faq:
          type: array
          items:
            type: object
            required: [question, answer]
            properties:
              question: { type: string }
              answer: { type: string }
        ctaText: { type: string }
        suggestedPriceBand: { type: string }
...